Refacto AbstractPerimeterVoter
