+++ Backoffice/Security/Authorization/Voter/NodeVoter.php
@@ -2,7 +2,6 @@
 
 namespace OpenOrchestra\Backoffice\Security\Authorization\Voter;
 
-use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
 use OpenOrchestra\ModelInterface\Model\NodeInterface;
 use OpenOrchestra\UserBundle\Model\UserInterface;
 use OpenOrchestra\Backoffice\Security\ContributionActionInterface;
@@ -24,42 +23,16 @@ class NodeVoter extends AbstractPerimeterVoter
     }
 
     /**
-     * @return array
-     */
-    protected function getSupportedAttributes()
-    {
-        return array(
-            ContributionActionInterface::READ,
-            ContributionActionInterface::ADD,
-            ContributionActionInterface::EDIT,
-            ContributionActionInterface::DELETE
-        );
-    }
-
-    /**
-     * @param string         $attribute
-     * @param mixed          $subject
-     * @param TokenInterface $token
+     * Vote for Read action
+     *
+     * @param NodeInterface $node
+     * @param UserInterface $user
      *
      * @return bool
      */
-    protected function voteOnAttribute($attribute, $subject, TokenInterface $token)
+    protected function voteForReadAction($node, $user)
     {
-        $user = $token->getUser();
-
-        if ($this->isSuperAdmin($user)) {
-            return true;
-        }
-
-        if (ContributionActionInterface::READ == $attribute) {
-            return true;
-        }
-
-        if ($subject->getCreatedBy() == $user->getUsername()) {
-            return $this->voteForOwnedNode($attribute, $subject, $user);
-        }
-
-        return $this->voteForSomeoneElseNode($attribute, $subject, $user);
+        return true;
     }
 
     /**
@@ -71,7 +44,7 @@ class NodeVoter extends AbstractPerimeterVoter
      *
      * @return bool
      */
-    protected function voteForOwnedNode($action, NodeInterface $node, UserInterface $user)
+    protected function voteForOwnedSubject($action, $node, UserInterface $user)
     {
         return $user->hasRole(ContributionRoleInterface::NODE_CONTRIBUTOR)
             && $this->isSubjectInAllowedPerimeter($node->getPath(), $user, NodeInterface::ENTITY_TYPE);
@@ -86,7 +59,7 @@ class NodeVoter extends AbstractPerimeterVoter
      *
      * @return bool
      */
-    protected function voteForSomeoneElseNode($action, NodeInterface $node, UserInterface $user)
+    protected function voteForSomeoneElseSubject($action, $node, UserInterface $user)
     {
         $requiredRole = ContributionRoleInterface::NODE_CONTRIBUTOR;
